pyproject.toml
[tool.poetry]
name = "dextr"
version = "0.1.0"
description = ""
authors = ["lazy-geeek <lazy-geek@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2024.1"
websockets = "^12.0"
termcolor = "^2.4.0"
blessed = "^1.20.0"
blessings = "^1.7"
python-decouple = "^3.8"
ccxt = "^4.3.6"
schedule = "^1.2.1"
hyperliquid-python-sdk = "^0.3.0"
pandas = "^2.2.2"
ta-lib = "^0.4.28"
pandas-ta = "^0.3.14b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
